# packages/api-contracts/openapi/wishpool.yaml
openapi: 3.1.0
info:
  title: WishPool API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    WishPool core API contract for mobile apps, parent web, admin web, and backend services.
    All write endpoints are command-oriented and should be called with an Idempotency-Key header.
servers:
  - url: http://localhost:8080
    description: Local core-api
security:
  - bearerAuth: []
...
paths:
...
  /families:
    post:
      tags: [Families]
      operationId: createFamily
      summary: Create a family.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFamilyRequest'
      responses:
        '201':
          description: Family created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        default:
          $ref: '#/components/responses/Error'
  /families/{familyId}:
    get:
      tags: [Families]
      operationId: getFamily
      summary: Get family details.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: Family.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        default:
          $ref: '#/components/responses/Error'
  /families/{familyId}/members:
    get:
      tags: [Families]
      operationId: listFamilyMembers
      summary: List family members.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: Family members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyMember'
        default:
          $ref: '#/components/responses/Error'
  /families/{familyId}/invites:
    post:
      tags: [Families]
      operationId: inviteParent
      summary: Invite a parent to a family.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteParentRequest'
      responses:
        '201':
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyInvite'
        default:
          $ref: '#/components/responses/Error'
  /families/{familyId}/children:
    get:
      tags: [Children]
      operationId: listChildren
      summary: List children in a family.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: Children.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChildProfile'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Children]
      operationId: createChild
      summary: Create a child profile.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        '201':
          description: Child profile created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
        default:
          $ref: '#/components/responses/Error'
  /children/{childId}:
    patch:
      tags: [Children]
      operationId: updateChild
      summary: Update child profile.
      parameters:
        - $ref: '#/components/parameters/ChildId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChildRequest'
      responses:
        '200':
          description: Child profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
        default:
          $ref: '#/components/responses/Error'
...
  /memories:
    get:
      tags: [Memories]
      operationId: listMemories
      summary: List memories for a child.
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Memory timeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryTimeline'
        default:
          $ref: '#/components/responses/Error'
  /memories/{memoryId}:
    get:
      tags: [Memories]
      operationId: getMemory
      summary: Get a memory detail.
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Weekly memory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyMemory'
        default:
          $ref: '#/components/responses/Error'
  /memories/{memoryId}/export:
    post:
      tags: [Memories]
      operationId: exportMemory
      summary: Request memory export.
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportMemoryRequest'
      responses:
        '202':
          description: Export requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryExport'
        default:
          $ref: '#/components/responses/Error'
...
  /internal/admin/dashboard:
    get:
      tags: [Admin]
      security:
        - internalToken: []
      operationId: getAdminDashboard
      summary: Get local administration dashboard counters.
      responses:
        '200':
          description: Dashboard counters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboard'
        default:
          $ref: '#/components/responses/Error'
  /internal/admin/families:
    get:
      tags: [Admin]
      security:
        - internalToken: []
      operationId: listAdminFamilies
      summary: List family metadata for administration.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, locked, deleting, deleted]
      responses:
        '200':
          description: Family metadata.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminFamilySummary'
        default:
          $ref: '#/components/responses/Error'
  /internal/admin/privacy-requests:
    get:
      tags: [Admin]
      security:
        - internalToken: []
      operationId: listAdminPrivacyRequests
      summary: List privacy requests for administration.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Privacy request rows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminPrivacyRequest'
        default:
          $ref: '#/components/responses/Error'
  /internal/admin/audit-logs:
    get:
      tags: [Admin]
      security:
        - internalToken: []
      operationId: listAdminAuditLogs
      summary: List audit logs for administration.
      parameters:
        - name: familyId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Audit log rows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAuditLog'
        default:
          $ref: '#/components/responses/Error'
  /internal/admin/media-access-grants:
    post:
      tags: [Admin]
      security:
        - internalToken: []
      operationId: grantAdminMediaAccess
      summary: Create an audited short-lived media access grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMediaAccessGrantRequest'
      responses:
        '200':
          description: Media access grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMediaAccessGrant'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
...
    Family:
      type: object
      required: [id, name, timezone, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        status:
          type: string
          enum: [active, locked, deleting, deleted]
    FamilyMember:
      type: object
      required: [id, familyId, userId, role, displayName, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/MemberRole'
        childId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        displayName:
          type: string
        status:
          type: string
          enum: [active, invited, suspended, removed]
...
    MemberRole:
      type: string
      enum: [parent_owner, parent, child_device]
    CreateFamilyRequest:
      type: object
      required: [name, timezone]
      properties:
        name:
          type: string
          minLength: 1
        timezone:
          type: string
          default: Asia/Shanghai
        firstChild:
          $ref: '#/components/schemas/CreateChildRequest'
    InviteParentRequest:
      type: object
      required: [contact]
      properties:
        contact:
          type: string
          description: Phone number for the self-hosted version.
    FamilyInvite:
      type: object
      required: [id, familyId, status, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted, expired, revoked]
        expiresAt:
          type: string
          format: date-time
    CreateChildRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
        birthYear:
          type: integer
          minimum: 2000
          maximum: 2100
        avatarAsset:
          type: string
        roomTheme:
          type: string
          default: star_cabin
    UpdateChildRequest:
      type: object
      properties:
        nickname:
          type: string
        birthYear:
          type: integer
        avatarAsset:
          type: string
        roomTheme:
          type: string
    ChildProfile:
      type: object
      required: [id, familyId, nickname, roomTheme, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        nickname:
          type: string
        birthYear:
          anyOf:
            - type: integer
            - type: 'null'
        avatarAsset:
          anyOf:
            - type: string
            - type: 'null'
        roomTheme:
          type: string
        status:
          type: string
          enum: [active, archived]
...
    MediaAsset:
      type: object
      required: [id, familyId, purpose, storageKey, contentType, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        purpose:
          type: string
        storageKey:
          type: string
        contentType:
          type: string
        status:
          type: string
          enum: [upload_pending, uploaded, processing, ready, failed, deleted]
        downloadUrl:
          anyOf:
            - type: string
              format: uri
            - type: 'null'
...
    MemoryTimeline:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyMemory'
        nextCursor:
          anyOf:
            - type: string
            - type: 'null'
    WeeklyMemory:
      type: object
      required: [id, childId, weekId, title, status]
      properties:
        id:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        title:
          type: string
        summary:
          type: object
          additionalProperties: true
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        status:
          type: string
          enum: [generating, generated, exported]
    ExportMemoryRequest:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [pdf, long_image]
    MemoryExport:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [requested, processing, ready, failed]
        media:
          $ref: '#/components/schemas/MediaAsset'
...
    AdminDashboard:
      type: object
      required: [generatedAt, familyCount, activeChildCount, pendingReviewCount, pendingNotificationCount, pendingOutboxCount, processingMediaCount, runningAiJobCount, openPrivacyRequestCount]
      properties:
        generatedAt:
          type: string
          format: date-time
        familyCount:
          type: integer
          format: int64
        activeChildCount:
          type: integer
          format: int64
        pendingReviewCount:
          type: integer
          format: int64
        pendingNotificationCount:
          type: integer
          format: int64
        pendingOutboxCount:
          type: integer
          format: int64
        processingMediaCount:
          type: integer
          format: int64
        runningAiJobCount:
          type: integer
          format: int64
        openPrivacyRequestCount:
          type: integer
          format: int64
    AdminFamilySummary:
      type: object
      required: [id, name, timezone, status, childCount, memberCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        status:
          type: string
          enum: [active, locked, deleting, deleted]
        childCount:
          type: integer
          format: int64
        memberCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    AdminPrivacyRequest:
      type: object
      required: [id, familyId, requestType, status, requestedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        requestType:
          type: string
          enum: [export, delete]
        status:
          type: string
          enum: [requested, verifying, locking_family, deleting_records, deleting_objects, verifying_deletion, completed, failed_needs_attention]
        requestedBy:
          type: string
          format: uuid
        exportMediaId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        reason:
          anyOf:
            - type: string
            - type: 'null'
        createdAt:
          type: string
          format: date-time
        completedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        errorMessage:
          anyOf:
            - type: string
            - type: 'null'
    AdminAuditLog:
      type: object
      required: [id, actorRole, action, resourceType, metadata, createdAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        actorUserId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        actorRole:
          type: string
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    AdminMediaAccessGrantRequest:
      type: object
      required: [familyId, mediaAssetId, reason]
      properties:
        familyId:
          type: string
          format: uuid
        mediaAssetId:
          type: string
          format: uuid
        reason:
          type: string
        expiresInMinutes:
          type: integer
          default: 30
          minimum: 5
          maximum: 120
    AdminMediaAccessGrant:
      type: object
      required: [mediaAssetId, familyId, accessUrl, expiresAt, auditLogId]
      properties:
        mediaAssetId:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        accessUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        auditLogId:
          type: string
          format: uuid
...